refactor redux to @reduxjs/toolkit
