+++ src/app/Old-store.js
@@ -1,0 +1,16 @@
+// refactored using Redux Toolkit, this is the old version renamed to Old-store.js
+
+/**
+    import { createStore, combineReducers } from "redux";
+    import { favoriteRecipesReducer } from "../features/favoriteRecipes/favoriteRecipesSlice.js";
+    import { searchTermReducer } from "../features/searchTerm/searchTermSlice.js";
+    import { allRecipesReducer } from "../features/allRecipes/allRecipesSlice.js";
+    
+    export const store = createStore(
+      combineReducers({
+        favoriteRecipes: favoriteRecipesReducer,
+        searchTerm: searchTermReducer,
+        allRecipes: allRecipesReducer,
+      })
+    );
+*/
