+++ src/app/store.js
@@ -1,12 +1,12 @@
-import { createStore, combineReducers } from "redux";
-import { favoriteRecipesReducer } from "../features/favoriteRecipes/favoriteRecipesSlice.js";
+import { configureStore } from "@reduxjs/toolkit";
+import { favoriteRecipesSlice } from "../features/favoriteRecipes/favoriteRecipesSlice.js";
 import { searchTermReducer } from "../features/searchTerm/searchTermSlice.js";
 import { allRecipesReducer } from "../features/allRecipes/allRecipesSlice.js";
 
-export const store = createStore(
-  combineReducers({
-    favoriteRecipes: favoriteRecipesReducer,
+export const store = configureStore({
+  reducer: {
+    favoriteRecipes: favoriteRecipesSlice.reducer,
     searchTerm: searchTermReducer,
     allRecipes: allRecipesReducer,
-  })
-);
+  },
+});
