+++ src/features/favoriteRecipes/Old-favoriteRecipesSlice.js
@@ -1,0 +1,47 @@
+// refactored using Redux Toolkit, this is the old version renamed to Old-favoriteRecipesSlice.js
+
+/**
+  import { selectSearchTerm } from "../searchTerm/searchTermSlice.js";
+  
+  const initialState = [];
+  export const favoriteRecipesReducer = (
+    favoriteRecipes = initialState,
+    action
+  ) => {
+    switch (action.type) {
+      case "favoriteRecipes/addRecipe":
+        return [...favoriteRecipes, action.payload];
+      case "favoriteRecipes/removeRecipe":
+        return favoriteRecipes.filter(
+          (recipe) => recipe.id !== action.payload.id
+        );
+      default:
+        return favoriteRecipes;
+    }
+  };
+  
+  export function addRecipe(recipe) {
+    return {
+      type: "favoriteRecipes/addRecipe",
+      payload: recipe,
+    };
+  }
+  
+  export function removeRecipe(recipe) {
+    return {
+      type: "favoriteRecipes/removeRecipe",
+      payload: recipe,
+    };
+  }
+  
+  export const selectFavoriteRecipes = (state) => state.favoriteRecipes;
+  
+  export const selectFilteredFavoriteRecipes = (state) => {
+    const favoriteRecipes = selectFavoriteRecipes(state);
+    const searchTerm = selectSearchTerm(state);
+  
+    return favoriteRecipes.filter((recipe) =>
+      recipe.name.toLowerCase().includes(searchTerm.toLowerCase())
+    );
+  };
+*/
