+++ src/features/favoriteRecipes/favoriteRecipesSlice.js
@@ -1,35 +1,18 @@
+import { createSlice } from "@reduxjs/toolkit";
 import { selectSearchTerm } from "../searchTerm/searchTermSlice.js";
 
-const initialState = [];
-export const favoriteRecipesReducer = (
-  favoriteRecipes = initialState,
-  action
-) => {
-  switch (action.type) {
-    case "favoriteRecipes/addRecipe":
-      return [...favoriteRecipes, action.payload];
-    case "favoriteRecipes/removeRecipe":
-      return favoriteRecipes.filter(
-        (recipe) => recipe.id !== action.payload.id
-      );
-    default:
-      return favoriteRecipes;
-  }
-};
-
-export function addRecipe(recipe) {
-  return {
-    type: "favoriteRecipes/addRecipe",
-    payload: recipe,
-  };
-}
-
-export function removeRecipe(recipe) {
-  return {
-    type: "favoriteRecipes/removeRecipe",
-    payload: recipe,
-  };
-}
+export const favoriteRecipesSlice = createSlice({
+  name: "favoriteRecipes",
+  initialState: [],
+  reducers: {
+    addRecipe: (state, action) => {
+      return [...state, action.payload];
+    },
+    removeRecipe: (state, action) => {
+      return state.filter((recipe) => recipe.id !== action.payload.id);
+    },
+  },
+});
 
 export const selectFavoriteRecipes = (state) => state.favoriteRecipes;
 
@@ -41,3 +24,6 @@ export const selectFilteredFavoriteRecipes = (state) => {
     recipe.name.toLowerCase().includes(searchTerm.toLowerCase())
   );
 };
+
+export const { addRecipe, removeRecipe } = favoriteRecipesSlice.actions;
+export default favoriteRecipesSlice.reducer;
